python/requirements.txt
# Core dependencies
pyjwt>=2.8.0
requests>=2.31.0
stackit-stackitmarketplace
stackit-core